Create widgetmanagers
