+++ pygame_menu/widgets/widget/dropselect_multiple.py
@@ -10,6 +10,7 @@ __all__ = [
 
     # Main Class
     'DropSelectMultiple',
+    'DropSelectMultipleManager',
 
     # Constants
     'DROPSELECT_MULTIPLE_SFORMAT_LIST_COMMA',
@@ -22,10 +23,13 @@ __all__ = [
 ]
 
 import pygame
+import pygame_menu
 
+from abc import ABC
 from pygame_menu.font import FontType
 from pygame_menu.locals import POSITION_NORTHWEST, POSITION_SOUTHEAST
 from pygame_menu.utils import assert_color, assert_vector, is_callable
+from pygame_menu.widgets.core.widget import AbstractWidgetManager, Widget
 from pygame_menu.widgets.widget.button import Button
 from pygame_menu.widgets.widget.dropselect import DropSelect
 
@@ -531,3 +535,284 @@ class DropSelectMultiple(DropSelect):
 
     def change(self, *args) -> Any:
         super(DropSelectMultiple, self).change()
+
+
+class DropSelectMultipleManager(AbstractWidgetManager, ABC):
+    """
+    DropSelectMultiple manager.
+    """
+
+    def dropselect_multiple(
+            self,
+            title: Any,
+            items: Union[List[Tuple[Any, ...]], List[str]],
+            default: Optional[Union[int, List[int]]] = None,
+            dropselect_multiple_id: str = '',
+            max_selected: int = 0,
+            onchange: CallbackType = None,
+            onreturn: CallbackType = None,
+            onselect: Optional[Callable[[bool, 'Widget', 'pygame_menu.Menu'], Any]] = None,
+            open_middle: bool = False,
+            placeholder: str = 'Select an option',
+            placeholder_add_to_selection_box: bool = True,
+            placeholder_selected: str = '{0} selected',
+            selection_placeholder_format: DropSelectMultipleSFormatType = DROPSELECT_MULTIPLE_SFORMAT_TOTAL,
+            **kwargs
+    ) -> 'pygame_menu.widgets.DropSelectMultiple':
+        """
+        Add a dropselect multiple to the Menu: Drop select multiple is a drop
+        select which can select many options at the same time. This drops a
+        vertical frame if requested.
+
+        The items of the DropSelectMultiple are:
+
+        .. code-block:: python
+
+            items = [('Item1', a, b, c...), ('Item2', d, e, f...), ('Item3', g, h, i...)]
+
+        The callbacks receive the current selected items (tuple) and the indices
+        (tuple), where ``selected_item=widget.get_value()`` and
+        ``selected_index=widget.get_index()``:
+
+        .. code-block:: python
+
+            onchange((selected_item, selected_index), **kwargs)
+            onreturn((selected_item, selected_index), **kwargs)
+
+        For example, if ``selected_index=[0, 2]`` then ``selected_item=[('Item1', a, b, c...), ('Item3', g, h, i...)]``.
+
+        If ``onselect`` is defined, the callback is executed as follows, where
+        ``selected`` is a boolean representing the selected status:
+
+        .. code-block:: python
+
+            onselect(selected, widget, menu)
+
+        kwargs (Optional)
+            - ``align``                         (str) – Widget `alignment <https://pygame-menu.readthedocs.io/en/latest/_source/themes.html#alignment>`_
+            - ``background_color``              (tuple, list, str, int, :py:class:`pygame.Color`, :py:class:`pygame_menu.baseimage.BaseImage`) – Color of the background. ``None`` for no-color
+            - ``background_inflate``            (tuple, list) – Inflate background on x-axis and y-axis (x, y) in px
+            - ``border_color``                  (tuple, list, str, int, :py:class:`pygame.Color`) – Widget border color. ``None`` for no-color
+            - ``border_inflate``                (tuple, list) – Widget border inflate on x-axis and y-axis (x, y) in px
+            - ``border_position``               (str, tuple, list) – Widget border positioning. It can be a single position, or a tuple/list of positions. Only are accepted: north, south, east, and west. See :py:mod:`pygame_menu.locals`
+            - ``border_width``                  (int) – Border width in px. If ``0`` disables the border
+            - ``cursor``                        (int, :py:class:`pygame.cursors.Cursor`, None) – Cursor of the widget if the mouse is placed over
+            - ``float``                         (bool) - If ``True`` the widget don't contributes width/height to the Menu widget positioning computation, and don't add one unit to the rows
+            - ``float_origin_position``         (bool) - If ``True`` the widget position is set to the top-left position of the Menu if the widget is floating
+            - ``font_background_color``         (tuple, list, str, int, :py:class:`pygame.Color`, None) – Widget font background color
+            - ``font_color``                    (tuple, list, str, int, :py:class:`pygame.Color`) – Widget font color
+            - ``font_name``                     (str, :py:class:`pathlib.Path`, :py:class:`pygame.font.Font`) – Widget font path
+            - ``font_shadow_color``             (tuple, list, str, int, :py:class:`pygame.Color`) – Font shadow color
+            - ``font_shadow_offset``            (int) – Font shadow offset in px
+            - ``font_shadow_position``          (str) – Font shadow position, see locals for position
+            - ``font_shadow``                   (bool) – Font shadow is enabled or disabled
+            - ``font_size``                     (int) – Font size of the widget
+            - ``margin``                        (tuple, list) – Widget (left, bottom) margin in px
+            - ``padding``                       (int, float, tuple, list) – Widget padding according to CSS rules. General shape: (top, right, bottom, left)
+            - ``readonly_color``                (tuple, list, str, int, :py:class:`pygame.Color`) – Color of the widget if readonly mode
+            - ``readonly_selected_color``       (tuple, list, str, int, :py:class:`pygame.Color`) – Color of the widget if readonly mode and is selected
+            - ``selection_color``               (tuple, list, str, int, :py:class:`pygame.Color`) – Color of the selected widget; only affects the font color
+            - ``selection_effect``              (:py:class:`pygame_menu.widgets.core.Selection`) – Widget selection effect
+            - ``tab_size``                      (int) – Width of a tab character
+
+        kwargs for modifying selection box/option style (Optional)
+            - ``scrollbar_color``                       (tuple, list, str, int, :py:class:`pygame.Color`) – Scrollbar color
+            - ``scrollbar_cursor``                      (int, :py:class:`pygame.cursors.Cursor`, None) – Cursor of the scrollbars if the mouse is placed over
+            - ``scrollbar_shadow_color``                (tuple, list, str, int, :py:class:`pygame.Color`) – Color of the shadow of each scrollbar
+            - ``scrollbar_shadow_offset``               (int) – Offset of the scrollbar shadow in px
+            - ``scrollbar_shadow_position``             (str) – Position of the scrollbar shadow. See :py:mod:`pygame_menu.locals`
+            - ``scrollbar_shadow``                      (bool) – Indicate if a shadow is drawn on each scrollbar
+            - ``scrollbar_slider_color``                (tuple, list, str, int, :py:class:`pygame.Color`) – Color of the sliders
+            - ``scrollbar_slider_hover_color``          (tuple, list, str, int, :py:class:`pygame.Color`) – Color of the slider if hovered or clicked
+            - ``scrollbar_slider_pad``                  (int, float) – Space between slider and scrollbars borders in px
+            - ``scrollbar_thick``                       (int) – Scrollbar thickness in px
+            - ``scrollbars``                            (str) – Scrollbar position. See :py:mod:`pygame_menu.locals`
+            - ``selection_box_arrow_color``             (tuple, list, str, int, :py:class:`pygame.Color`) – Selection box arrow color
+            - ``selection_box_arrow_margin``            (tuple) – Selection box arrow margin (left, right, vertical) in px
+            - ``selection_box_bgcolor``                 (tuple, list, str, int, :py:class:`pygame.Color`) – Selection box background color
+            - ``selection_box_border_color``            (tuple, list, str, int, :py:class:`pygame.Color`) – Selection box border color
+            - ``selection_box_border_width``            (int) – Selection box border width
+            - ``selection_box_height``                  (int) – Selection box height, counted as how many options are packed before showing scroll
+            - ``selection_box_inflate``                 (tuple) – Selection box inflate on x-axis and y-axis in px
+            - ``selection_box_margin``                  (tuple, list) – Selection box on x-axis and y-axis (x, y) margin from title in px
+            - ``selection_box_text_margin``             (int) – Selection box text margin (left) in px
+            - ``selection_box_width``                   (int) – Selection box width in px. If ``0`` compute automatically to fit placeholder
+            - ``selection_infinite``                    (bool) – If ``True`` selection can rotate through bottom/top
+            - ``selection_option_active_bgcolor``       (tuple, list, str, int, :py:class:`pygame.Color`) – Active option(s) background color; active options is the currently active (by user)
+            - ``selection_option_active_font_color``    (tuple, list, str, int, :py:class:`pygame.Color`) – Active option(s) font color
+            - ``selection_option_border_color``         (tuple, list, str, int, :py:class:`pygame.Color`) – Option border color
+            - ``selection_option_border_width``         (int) – Option border width
+            - ``selection_option_font_color``           (tuple, list, str, int, :py:class:`pygame.Color`) – Option font color
+            - ``selection_option_font_size``            (int, None) – Option font size. If ``None`` use the 75% of the widget font size
+            - ``selection_option_font``                 (str, :py:class:`pathlib.Path`, :py:class:`pygame.font.Font`) – Option font. If ``None`` use the same font as the widget
+            - ``selection_option_padding``              (int, float, tuple, list) – Selection padding. See padding styling
+            - ``selection_option_selected_bgcolor``     (tuple, list, str, int, :py:class:`pygame.Color`) – Selected option background color
+            - ``selection_option_selected_box_border``  (int) – Box border width in px
+            - ``selection_option_selected_box_color``   (tuple, list, str, int, :py:class:`pygame.Color`) – Box color
+            - ``selection_option_selected_box_height``  (int, float) – Height of the selection box relative to the options height
+            - ``selection_option_selected_box_margin``  (tuple, list) – Option box margin (left, right, vertical) in px
+            - ``selection_option_selected_box``         (bool) – Draws a box in the selected option(s)
+            - ``selection_option_selected_font_color``  (tuple, list, str, int, :py:class:`pygame.Color`) – Selected option font color
+
+        .. note::
+
+            All theme-related optional kwargs use the default Menu theme if not
+            defined.
+
+        .. note::
+
+            This is applied only to the base Menu (not the currently displayed,
+            stored in ``_current`` pointer); for such behaviour apply to
+            :py:meth:`pygame_menu.menu.Menu.get_current` object.
+
+        .. warning::
+
+            Be careful with kwargs collision. Consider that all optional documented
+            kwargs keys are removed from the object.
+
+        :param title: Drop select title
+        :param items: Item list of the drop select; format ``[('Item1', a, b, c...), ('Item2', d, e, f...)]``
+        :param default: Index(es) of default item(s) to display. If ``None`` no item is selected
+        :param dropselect_multiple_id: ID of the dropselect multiple
+        :param max_selected: Max items to be selected. If ``0`` there's no limit
+        :param onchange: Callback when changing the drop select item
+        :param onreturn: Callback when pressing return on the selected item
+        :param onselect: Function when selecting the widget
+        :param open_middle: If ``True`` the selection box is opened in the middle of the menu
+        :param placeholder: Text shown if no option is selected yet
+        :param placeholder_add_to_selection_box: If ``True`` adds the placeholder button to the selection box
+        :param placeholder_selected: Text shown if option is selected. Accepts the number of selected options
+        :param selection_placeholder_format: Format of the string replaced in ``placeholder_selected``. Can be a predefined string type ("total", "comma-list", "hyphen-list", or any other string which will join the list) or a function that receives the list of selected items and returns a string
+        :param kwargs: Optional keyword arguments
+        :return: Widget object
+        :rtype: :py:class:`pygame_menu.widgets.DropSelectMultiple`
+        """
+        # Filter widget attributes to avoid passing them to the callbacks
+        attributes = self._filter_widget_attributes(kwargs)
+
+        # Get selection box properties
+        selection_box_arrow_color = kwargs.pop('selection_box_arrow_color',
+                                               self._theme.widget_box_arrow_color)
+        selection_box_arrow_margin = kwargs.pop('selection_box_arrow_margin',
+                                                self._theme.widget_box_arrow_margin)
+        selection_box_bgcolor = kwargs.pop('selection_box_bgcolor',
+                                           self._theme.widget_box_background_color)
+        selection_box_border_color = kwargs.pop('selection_box_border_color',
+                                                self._theme.widget_box_border_color)
+        selection_box_border_width = kwargs.pop('selection_box_border_width',
+                                                self._theme.widget_box_border_width)
+        selection_box_height = kwargs.pop('selection_box_height', 3)
+        selection_box_inflate = kwargs.pop('selection_box_inflate',
+                                           self._theme.widget_border_inflate)
+        selection_box_margin = kwargs.pop('selection_box_margin',
+                                          self._theme.widget_box_margin)
+        selection_box_text_margin = kwargs.pop('selection_box_text_margin',
+                                               self._theme.widget_box_arrow_margin[0])
+        selection_box_width = kwargs.pop('selection_box_width', 0)
+        selection_infinite = kwargs.pop('selection_infinite', False)
+        selection_option_active_bgcolor = kwargs.pop('selection_option_active_bgcolor',
+                                                     (188, 227, 244))
+        selection_option_active_font_color = kwargs.pop('selection_option_active_font_color',
+                                                        (0, 0, 0))
+        selection_option_border_color = kwargs.pop('selection_option_border_color',
+                                                   self._theme.scrollbar_color)
+        selection_option_border_width = kwargs.pop('selection_option_border_width',
+                                                   self._theme.widget_box_border_width)
+        # selection_option_cursor = kwargs.pop('selection_option_cursor', None)
+        selection_option_font = kwargs.pop('selection_option_font', None)
+        selection_option_font_color = kwargs.pop('selection_option_font_color', (0, 0, 0))
+        selection_option_font_size = kwargs.pop('selection_option_font_size', None)
+        selection_option_padding = kwargs.pop('selection_option_padding', (2, 5))
+        selection_option_selected_bgcolor = kwargs.pop('selection_option_selected_bgcolor',
+                                                       (142, 247, 141))
+        selection_option_selected_box = kwargs.pop('selection_option_selected_box', True)
+        selection_option_selected_box_border = kwargs.pop('selection_option_selected_box_border',
+                                                          self._theme.widget_box_border_width)
+        selection_option_selected_box_color = kwargs.pop('selection_option_selected_box_color',
+                                                         self._theme.widget_box_arrow_color)
+        selection_option_selected_box_height = kwargs.pop('selection_option_selected_box_height',
+                                                          0.5)
+        selection_option_selected_box_margin = kwargs.pop('selection_option_selected_box_margin',
+                                                          (0, self._theme.widget_box_arrow_margin[1],
+                                                           self._theme.widget_box_arrow_margin[2]))
+        selection_option_selected_font_color = kwargs.pop('selection_option_selected_font_color',
+                                                          (0, 0, 0))
+
+        # Get selection box scrollbar properties
+        scrollbar_color = kwargs.pop('scrollbar_color', self._theme.scrollbar_color)
+        scrollbar_cursor = kwargs.pop('scrollbar_cursor', self._theme.scrollbar_cursor)
+        scrollbar_shadow_color = kwargs.pop('scrollbar_shadow_color',
+                                            self._theme.scrollbar_shadow_color)
+        scrollbar_shadow_offset = kwargs.pop('scrollbar_shadow_offset',
+                                             self._theme.scrollbar_shadow_offset)
+        scrollbar_shadow_position = kwargs.pop('scrollbar_shadow_position',
+                                               self._theme.scrollbar_shadow_position)
+        scrollbar_shadow = kwargs.pop('scrollbar_shadow', self._theme.scrollbar_shadow)
+        scrollbar_slider_color = kwargs.pop('scrollbar_slider_color',
+                                            self._theme.scrollbar_slider_color)
+        scrollbar_slider_hover_color = kwargs.pop('scrollbar_slider_hover_color',
+                                                  self._theme.scrollbar_slider_hover_color)
+        scrollbar_slider_pad = kwargs.pop('scrollbar_slider_pad',
+                                          self._theme.scrollbar_slider_pad)
+        scrollbar_thick = kwargs.pop('scrollbar_thick', self._theme.scrollbar_thick)
+        scrollbars = kwargs.pop('scrollbars', self._theme.scrollarea_position)
+
+        widget = DropSelectMultiple(
+            default=default,
+            dropselect_id=dropselect_multiple_id,
+            items=items,
+            max_selected=max_selected,
+            onchange=onchange,
+            onreturn=onreturn,
+            onselect=onselect,
+            open_middle=open_middle,
+            placeholder=placeholder,
+            placeholder_add_to_selection_box=placeholder_add_to_selection_box,
+            placeholder_selected=placeholder_selected,
+            scrollbar_color=scrollbar_color,
+            scrollbar_cursor=scrollbar_cursor,
+            scrollbar_shadow=scrollbar_shadow,
+            scrollbar_shadow_color=scrollbar_shadow_color,
+            scrollbar_shadow_offset=scrollbar_shadow_offset,
+            scrollbar_shadow_position=scrollbar_shadow_position,
+            scrollbar_slider_color=scrollbar_slider_color,
+            scrollbar_slider_hover_color=scrollbar_slider_hover_color,
+            scrollbar_slider_pad=scrollbar_slider_pad,
+            scrollbar_thick=scrollbar_thick,
+            scrollbars=scrollbars,
+            selection_box_arrow_color=selection_box_arrow_color,
+            selection_box_arrow_margin=selection_box_arrow_margin,
+            selection_box_bgcolor=selection_box_bgcolor,
+            selection_box_border_color=selection_box_border_color,
+            selection_box_border_width=selection_box_border_width,
+            selection_box_height=selection_box_height,
+            selection_box_inflate=selection_box_inflate,
+            selection_box_margin=selection_box_margin,
+            selection_box_text_margin=selection_box_text_margin,
+            selection_box_width=selection_box_width,
+            selection_infinite=selection_infinite,
+            selection_option_active_bgcolor=selection_option_active_bgcolor,
+            selection_option_active_font_color=selection_option_active_font_color,
+            selection_option_border_color=selection_option_border_color,
+            selection_option_border_width=selection_option_border_width,
+            # selection_option_cursor=selection_option_cursor,
+            selection_option_font=selection_option_font,
+            selection_option_font_color=selection_option_font_color,
+            selection_option_font_size=selection_option_font_size,
+            selection_option_padding=selection_option_padding,
+            selection_option_selected_bgcolor=selection_option_selected_bgcolor,
+            selection_option_selected_box=selection_option_selected_box,
+            selection_option_selected_box_border=selection_option_selected_box_border,
+            selection_option_selected_box_color=selection_option_selected_box_color,
+            selection_option_selected_box_height=selection_option_selected_box_height,
+            selection_option_selected_box_margin=selection_option_selected_box_margin,
+            selection_option_selected_font_color=selection_option_selected_font_color,
+            selection_placeholder_format=selection_placeholder_format,
+            title=title,
+            **kwargs
+        )
+
+        self._configure_widget(widget=widget, **attributes)
+        self._append_widget(widget)
+
+        return widget
